Fix download link not working in Edge and Chrome
Previously, the `src` attribute of the image previews are using object
URLs created from the selected image, which are immediately revoked when
the images are loaded on the DOM. This breaks the download link because
the blob that is being referenced by its `href` attribute has already
been cleared from memory.

This commit doesn't revoke the object URL on image load, but only when a
new image is being compressed.

The `loadImage()` function is also refactored so it only runs
`canvas.toBlob()` when loading the compressed version of an image. When
loading the original version of an image, it skips drawing onto the
canvas and simply sets the `src` attribute of the `image-1` slot using
the selected file's data URL.