Remove click handling from <image-card>
Since the image queue now handles card selection events, we can remove
event listeners from the image card. The button in the filename now only
acts as a way to manage focus. Alternatively, we can remove the button
completely and set `tabindex="0"` to the image card itself if it's a
better focus management implementation.