Bring back click handling logic to <image-card>
Putting the logic inside the <image-card> should make it easier to
prevent card selection when the card is disabled or loading. For
example, when an image card is still compressing, we have nothing to
show in the full image preview yet, so the user should not be able to
select that image card.