Clean up component styles
Since I refactored the components, a lot of CSS styles are not needed
anymore.

- img-Input.css -> not needed since we switched to a behavior-only
  <drop-zone> that treats the whole page as the drop zone

- sidebar.css -> replaced by the sidebar utility class

- slider.css -> slider styles are colocated in quality-slider.webc

- text-area.css -> we don't need textarea anymore

The styles in `button.css` have been cleaned so that it only handles
button styles, not the layout or context in which the buttons may
appear.

Same for `image-compare.css` where I removed most lines, only adjusting
the focus width and color.