Fix image preview bug
This commit fixes a bug where the image preview reverts to a previously
selected image that just finished compressing after a quality change.

For example, let's say I have two images on queue: one PNG and one JPEG.
The PNG file is over 100 KB, so I lower the quality slider. PNGs take a
while to compress, so I click on the other image to inspect it. After a
few seconds, the <image-preview> will update to show the PNG image, even
though the JPEG image is the one currently selected.

To fix this, I updated the `qualitychange` handler to update the image
preview only if the user did not select a different image card.