Add bump-version workflow
This commit adds a workflow that triggers when a pull request is merged
to the main branch. By default, it will bump the `minor` version and
create a corresponding tag. For example, if the old tag is `v1.0.0` and
a PR is merged to main, the merge commit will be tagged `v1.1.0`.

To bump the `major` version, include the tag `#major` in the merge
commit message; to bump the `patch` version, include the tag `#patch` in
the merge commit message.

When merging hotfixes, include `#patch` in the merge commit message to
bump the patch version and reflect the proper semantic versioning.

See https://github.com/anothrNick/github-tag-action for more info.